+++ pyasn1/codec/native/decoder.py
@@ -34,7 +34,7 @@ class SequenceOrSetPayloadDecoder(object):
     def __call__(self, pyObject, asn1Spec, decodeFun=None, **options):
         asn1Value = asn1Spec.clone()
 
-        componentsTypes = asn1Spec.componentType
+        componentsTypes = asn1Value.componentType
 
         for field in asn1Value:
             if field in pyObject:
@@ -48,7 +48,7 @@ class SequenceOfOrSetOfPayloadDecoder(object):
         asn1Value = asn1Spec.clone()
 
         for pyValue in pyObject:
-            asn1Value.append(decodeFun(pyValue, asn1Spec.componentType), **options)
+            asn1Value.append(decodeFun(pyValue, asn1Value.componentType), **options)
 
         return asn1Value
 
@@ -57,7 +57,7 @@ class ChoicePayloadDecoder(object):
     def __call__(self, pyObject, asn1Spec, decodeFun=None, **options):
         asn1Value = asn1Spec.clone()
 
-        componentsTypes = asn1Spec.componentType
+        componentsTypes = asn1Value.componentType
 
         for field in pyObject:
             if field in componentsTypes: